[rename] 파일위치 변경 (#99)
